Reformat

+++ qt_HMI/mainwindow.h
@@ -17,34 +17,34 @@ QT_END_NAMESPACE
 
 
 typedef enum {
-  LIGHTS_OFF_MOTOR = 0,
-  LIGHTS_ON = 1,
-  LIGHTS_DELAY = 2,
-  LIGHTS_OFF_SENSOR = 3
+    LIGHTS_OFF_MOTOR = 0,
+    LIGHTS_ON = 1,
+    LIGHTS_DELAY = 2,
+    LIGHTS_OFF_SENSOR = 3
 } LightsMode;
 
 struct State {
-  int voltage; // VOLT * 10
-  int sensor;  // Analog data - 0:1023 (1-1024?)
-  bool motor_running;
-  // char *ssid; // Wireless network ID
-  QString ssid_string;
-  LightsMode lights_mode = LIGHTS_OFF_MOTOR;
-  unsigned long int motor_start_millis;
+    double voltage;
+    int sensor;  // Analog data - 0:1023 (1-1024?)
+    bool motor_running;
+    // char *ssid; // Wireless network ID
+    QString ssid_string;
+    LightsMode lights_mode = LIGHTS_OFF_MOTOR;
+    unsigned long int motor_start_millis;
 };
 
 struct Settings {
-  double motor_running_voltage_threshold = 0;
-  int delay_lights_millis =
-      3000; // Delay to turn lights on after the motor is running
-  int sensor_threshold_to_turn_on = 0;
-  int sensor_threshold_to_turn_off = 0;
-  int wifi_connect_timeout =
-      15000; // Timeout connectiong to WiFi and create self AP
-  bool winter_mode = false;
-  bool use_bufferd_udp = false;
-  int localUdpPort = 0;
-  int remoteUdpPort = 0;
+    double motor_running_voltage_threshold = 0;
+    int delay_lights_millis =
+            3000; // Delay to turn lights on after the motor is running
+    int sensor_threshold_to_turn_on = 0;
+    int sensor_threshold_to_turn_off = 0;
+    int wifi_connect_timeout =
+            15000; // Timeout connectiong to WiFi and create self AP
+    bool winter_mode = false;
+    bool use_bufferd_udp = false;
+    int localUdpPort = 0;
+    int remoteUdpPort = 0;
 };
